+++ docs/HISTORY.rst
@@ -4,6 +4,8 @@ Know where your towel is
 1.0a2 (unreleased)
 ------------------
 
+- Add 'signature' field to behavior. [hvelarde]
+
 - Fix default value of fields to be an empty unicode string. [hvelarde]